fix(ci): override working-directory in build-summary job to use root directory
The build-summary job was failing because it inherited the workflow-level defaults.run.working-directory: browser-extension, but this job doesn't have a checkout step, so the directory doesn't exist. Since this job only checks the status of the previous build job and doesn't need to access any files, it should run in the root directory.

This fixes the error: 'Error occurred trying to start process /usr/bin/bash with working directory /home/runner/work/nsx/nsx/browser-extension. No such file or directory'